Moved MILC MG interface details into new files
